+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## vN.N.N
 * Adding `alt_text` field to image asset forms.
+* Add eme-form-block to Blocksmith.
+* cms crash to white screen after paste
 
 
 ## v0.13.3